Struct amethyst_test::prelude::SequencerState

pub struct SequencerState<T, E> where
    E: Send + Sync + 'static,  { /* fields omitted */ }

Pushes each State onto the Amethyst state stack in reverse order (LIFO).

This implementation does not override the Transition returned by the State that is pushed to. Furthermore, it always transitions to the next state in the stack, which means there is no "opt-out" of going through the stack.

Methods

impl<T, E> SequencerState<T, E> where
E: Send + Sync + 'static,

pub fn new(states: Vec<Box<dyn State<T, E>>>) -> Self

Constructs a new SequencerState.